Migrated API to Common API
