+++ helper/CommonAPI.ts
@@ -44,3 +44,17 @@ export const uploadAsset = async (
 
   return res.json();
 };
+
+
+//Fetch Content Types
+export const fetchAllContentTypes = async (jwt: string): Promise<any[]> => {
+  try {
+    const res = await axios.get("/api/get-content-types", {
+      headers: { Authorization: `Bearer ${jwt}` },
+    });
+    return res.data.content_types;
+  } catch (err) {
+    console.error("❌ Error fetching all content types:", err);
+    return [];
+  }
+};
