--- helper/GenerateContentAPI.ts
@@ -1,13 +1,0 @@
-import axios from "axios";
-
-export const fetchAllContentTypes = async (jwt: string): Promise<any[]> => {
-  try {
-    const res = await axios.get("/api/get-content-types", {
-      headers: { Authorization: `Bearer ${jwt}` },
-    });
-    return res.data.content_types;
-  } catch (err) {
-    console.error("❌ Error fetching all content types:", err);
-    return [];
-  }
-};
